quark_py/optimizer.py
import numpy as np
from typing import List, Tuple, Optional
from .circuit import QuantumCircuit
from .encoding import QubitEncoding

class QuantumOptimizer:
    def __init__(self, 
                 n_qubits: int,
                 n_iterations: int = 1000,
                 learning_rate: float = 0.01,
                 temperature: float = 1.0):
        self.n_qubits = n_qubits
        self.n_iterations = n_iterations
        self.learning_rate = learning_rate
        self.temperature = temperature
        self.circuit = QuantumCircuit(n_qubits)
        self.encoding = QubitEncoding(n_qubits)
        
    def quantum_annealing(self, hamiltonian: np.ndarray, initial_state: Optional[np.ndarray] = None) -> np.ndarray:
        # Start with uniform superposition if no initial state
        if initial_state is None:
            state = np.ones(2**self.n_qubits) / np.sqrt(2**self.n_qubits)
        else:
            state = initial_state / np.linalg.norm(initial_state)
            
        for t in range(self.n_iterations):
            current_temp = self.temperature * (1 - t/self.n_iterations)
            state = self.circuit.apply_hamiltonian(state, hamiltonian, current_temp)
            state = self.circuit.phase_estimation(state)
            state = self.circuit.quantum_interference(state)
            state /= np.linalg.norm(state)
            
        return state
    
    def optimize_consensus(self, transaction_graph: np.ndarray, stake_weights: np.ndarray) -> Tuple[np.ndarray, float]:
        quantum_state = self.encoding.encode_graph(transaction_graph, stake_weights)
        hamiltonian = self._build_consensus_hamiltonian(transaction_graph, stake_weights)
        optimized_state = self.quantum_annealing(hamiltonian, quantum_state)
        ordering = self.encoding.decode_to_ordering(optimized_state)
        confidence = self._calculate_confidence(optimized_state)
        return ordering, confidence
    
    def optimize_model(self, weights: np.ndarray, importance_scores: np.ndarray, 
                      compression_ratio: float) -> Tuple[np.ndarray, float]:
        quantum_state = self.encoding.encode_weights(weights, importance_scores)
        hamiltonian = self._build_compression_hamiltonian(weights.shape, compression_ratio)
        optimized_state = self.quantum_annealing(hamiltonian, quantum_state)
        compressed_weights = self.encoding.decode_to_weights(optimized_state, weights.shape)
        accuracy = self._estimate_accuracy(optimized_state, importance_scores)
        return compressed_weights, accuracy
    
    def _build_consensus_hamiltonian(self, transaction_graph: np.ndarray, stake_weights: np.ndarray) -> np.ndarray:
        n = len(transaction_graph)
        hamiltonian = np.zeros((2**self.n_qubits, 2**self.n_qubits), dtype=complex)
        
        # Add transaction dependencies
        for i in range(n):
            for j in range(n):
                if transaction_graph[i,j]:
                    hamiltonian += self._build_dependency_term(i, j, n)
        
        # Add stake weights
        for i in range(n):
            hamiltonian += stake_weights[i] * self._build_stake_term(i, n)
            
        return hamiltonian
    
    def _build_compression_hamiltonian(self, weight_shape: Tuple[int, ...], compression_ratio: float) -> np.ndarray:
        total_params = np.prod(weight_shape)
        target_params = int(total_params * compression_ratio)
        hamiltonian = np.zeros((2**self.n_qubits, 2**self.n_qubits), dtype=complex)
        hamiltonian += self._build_sparsity_term(target_params)
        hamiltonian += self._build_preservation_term(weight_shape)
        return hamiltonian